ml/requirements.txt
# Coconut Pest Detection - ML Dependencies

# Core ML
tensorflow>=2.15.0
keras>=3.0.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
opencv-python>=4.8.0
Pillow>=10.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Jupyter
jupyter>=1.0.0
notebook>=7.0.0
ipykernel>=6.25.0

# Model Evaluation
scikit-learn>=1.3.0

# Image Augmentation
albumentations>=1.3.0

# API (for serving model)
flask>=3.0.0
flask-cors>=4.0.0

# Utilities
tqdm>=4.65.0
python-dotenv>=1.0.0
